- Build tool dependencies updated 	- Updated to Scala 2.11.1 	- Updated to ScalaTest 2.2.0 	- Updated to SBT 0.13.5 	- Updated to SBT Scalariform 1.3.0 	- Updated to SBT Unidoc 0.3.1 	- Updated to SBT PGP 0.8.3 - Runtime dependencies updated 	- Now dependent on Scala XML 1.0.2 (XML support was separated from standard Scala library) 	- Updated to Spray JSON 1.2.6 	- Updated to Avro 1.7.6 	- Updated to SLF4J 1.7.7 	- Updated to Logback 1.1.2

+++ core/src/main/scala/com/gensler/scalavro/io/primitive/AvroFloatIO.scala
@@ -29,7 +29,7 @@ trait AvroFloatIO extends AvroPrimitiveTypeIO[Float] {
   // JSON ENCODING
   ////////////////////////////////////////////////////////////////////////////
 
-  def writePrimitiveJson(value: Float) = JsNumber(BigDecimal(value))
+  def writePrimitiveJson(value: Float) = JsNumber(value.toDouble)
 
   def readJson(json: JsValue) = Try {
     json match {
